+++ src/lib/components/cards/ExecCard.svelte
