+++ src/lib/components/cards/MissionVisionCard.svelte
