+++ src/lib/components/headers/FeaturedEvent.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import LinkButton from '$lib/components/LinkButton.svelte';
+    import LinkButton from '$lib/components/buttons/LinkButton.svelte';
     import type { Snippet } from 'svelte';
     import { Tag } from '$lib/data/events';
