+++ src/lib/components/miscellaneous/GeneralInfo.svelte
