+++ src/lib/components/nav/NavBar.svelte
@@ -10,7 +10,7 @@
     import { get } from '$lib/stores/color-scheme';
     import upcsi from '$lib/assets/icons/upcsi.svg';
 
-    import LinkButtonNavBar from '$lib/components/LinkButtonNavBar.svelte';
+    import LinkButtonNavBar from '$lib/components/buttons/LinkButtonNavBar.svelte';
 
     const theme = get();
     let mobileMenu = $state(false);
