+++ src/lib/components/panels/ContentContainer.svelte
