+++ src/lib/components/panels/SocialMediaPanel.svelte
