+++ src/lib/components/panels/TeamPanel.svelte
