+++ src/routes/about/+page.svelte
@@ -1,7 +1,7 @@
 <script>
-    import Contact from '$lib/components/Contact.svelte';
-    import MissionVision from './MissionVision.svelte';
-    import SectionHeaderTextContainer from './SectionHeaderTextContainer.svelte';
+    import Contact from '$lib/components/footers/Contact.svelte';
+    import MissionVision from '$lib/components/panels/MissionVisionPanel.svelte';
+    import SectionHeaderTextContainer from '../../lib/components/headers/SectionHeaderTextContainer.svelte';
 </script>
 
 <article class="prose max-w-full dark:prose-invert">
